Refactor Camera and Camera container, extracting section menu in separate file/container. Refactor camera selectors Integrate camera shooting menu into PlateSolving page
