+++ frontend/src/Navigation/NavbarMenuItems.js
@@ -7,10 +7,10 @@ import { Route } from "react-router-dom";
 import { SequencesListSectionMenu } from '../Sequences/SequencesList';
 import { SequenceSectionMenuContainer } from '../Sequences/SequenceContainer';
 import { SequenceJobImagesContainer } from '../SequenceJobs/ImagesContainer';
-import { CameraSectionMenuContainer } from '../Camera/CameraContainer';
 import { ImageSectionMenuContainer } from '../Image/ImageContainer';
 import { NavItem } from './NavbarMenu';
 import { PlateSolvingSectionMenuContainer } from '../PlateSolving/PlateSolvingContainer';
+import { CameraSectionMenu } from '../Camera/Camera';
 
 
 export const NavbarMenuItems = ({disabled, onClick = () => true}) => (
@@ -27,7 +27,7 @@ export const NavbarMenuItems = ({disabled, onClick = () => true}) => (
         <Route path={Routes.SEQUENCE_JOB_IMAGES} exact={true} render={
             ({match}) => <SequenceJobImagesContainer sequenceJob={match.params.itemId} />
         } />
-        <Route exact path={Routes.CAMERA_PAGE} component={CameraSectionMenuContainer} />
+        <Route exact path={Routes.CAMERA_PAGE} component={CameraSectionMenu} />
         <Route exact path={Routes.PLATE_SOLVING_PAGE} component={PlateSolvingSectionMenuContainer} />
         <Route path={Routes.IMAGE_PAGE} render={({match}) => <ImageSectionMenuContainer id={match.params.id} type={match.params.type} />} />
     </React.Fragment>
