+++ frontend/src/PlateSolving/PlateSolvingContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { PlateSolvingPage, PlateSolvingSectionMenu } from './PlateSolvingPage';
 import Actions from '../actions';
-import { plateSolvingContainerSelector } from './selectors';
+import { plateSolvingContainerSelector, plateSolvingSectionMenuSelector } from './selectors';
 
 const mapDispatchToProps = dispatch => ({
   setOption: (option, value) => dispatch(Actions.PlateSolving.setOption(option, value)),
@@ -14,8 +14,8 @@ export const PlateSolvingContainer = connect(
 )(PlateSolvingPage)
 
 export const PlateSolvingSectionMenuContainer = connect(
-  null,
-  mapDispatchToProps
+  plateSolvingSectionMenuSelector,
+  null, 
 )(PlateSolvingSectionMenu)
